replaced key types for repos by Long isntead of Integer

+++ src/main/java/repos/IDashboardRepository.java
@@ -3,5 +3,5 @@ package repos;
 import Models.Dashboard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IDashboardRepository extends JpaRepository<Dashboard, Integer> {
+public interface IDashboardRepository extends JpaRepository<Dashboard, Long> {
 }
